refactor(dao): DB Connection 부분 분리
