+++ src/main/java/chess/dao/setting/DBConnection.java
@@ -1,0 +1,46 @@
+package chess.dao.setting;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class DBConnection {
+    public Connection getConnection() {
+        final String server = "localhost:3306";                                   // MySQL 서버 주소
+        final String database = "mydb";                                           // MySQL DATABASE 이름
+        final String option = "?allowPublicKeyRetrieval=true&useSSL=false&serverTimezone=UTC";
+        final String userName = "da-nyee";                                        //  MySQL 서버 아이디
+        final String password = "1234";                                           // MySQL 서버 비밀번호
+
+        Connection con = null;
+
+        // 드라이버 로딩
+        try {
+            Class.forName("com.mysql.cj.jdbc.Driver");
+        } catch (ClassNotFoundException e) {
+            System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
+            e.printStackTrace();
+        }
+
+        // 드라이버 연결
+        try {
+            con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password);
+            System.out.println("정상적으로 연결되었습니다.");
+        } catch (SQLException e) {
+            System.err.println("연결 오류:" + e.getMessage());
+            e.printStackTrace();
+        }
+
+        return con;
+    }
+
+    // 드라이버 연결해제
+    public void closeConnection(Connection con) {
+        try {
+            if (con != null)
+                con.close();
+        } catch (SQLException e) {
+            System.err.println("con 오류:" + e.getMessage());
+        }
+    }
+}
